fix: change program live mothod

+++ openlist/openlist.sh
@@ -6,23 +6,31 @@ source $PROGRAM_PATH/config
 
 case "$1" in
 start)
-    if  ps -ef | grep $(cat $PROGRAM_PATH/$APP_NAME.pid) | grep -v grep > /dev/null ; then
-        echo "${APP_NAME} is runing..."
-        exit 0
+    if [ -f "$PROGRAM_PATH/$APP_NAME.pid" ]; then
+        if  ps -ef | grep $(cat $PROGRAM_PATH/$APP_NAME.pid) | grep -v grep > /dev/null ; then
+            echo "${APP_NAME} is runing..."
+            exit 0
+        fi
     fi
     cd $WORK_PATH
     ${APP_NAME} server > ${PROGRAM_PATH}/${APP_NAME}.log 2>&1 &
     echo $! > $PROGRAM_PATH/$APP_NAME.pid
     ;;
 stop)
-    if ps -ef | grep $(cat $PROGRAM_PATH/$APP_NAME.pid) | grep -v grep > /dev/null ; then
-        ps -ef | grep $(cat $PROGRAM_PATH/$APP_NAME.pid) | grep -v grep | awk '{print $2}' | xargs -I {} kill {}
-        rm $PROGRAM_PATH/$APP_NAME.pid
+    if [ -f "$PROGRAM_PATH/$APP_NAME.pid" ]; then
+        if  ps -ef | grep $(cat $PROGRAM_PATH/$APP_NAME.pid) | grep -v grep > /dev/null ; then
+            ps -ef | grep $(cat $PROGRAM_PATH/$APP_NAME.pid) | grep -v grep | awk '{print $2}' | xargs -I {} kill {}
+            rm $PROGRAM_PATH/$APP_NAME.pid
+        fi
     fi
     ;;
 status)
-    if ps -ef | grep $(cat $PROGRAM_PATH/$APP_NAME.pid) | grep -v grep > /dev/null ; then
-        echo "$APP_NAME is runing..."
+    if [ -f "$PROGRAM_PATH/$APP_NAME.pid" ]; then
+        if ps -ef | grep $(cat $PROGRAM_PATH/$APP_NAME.pid) | grep -v grep > /dev/null ; then
+            echo "$APP_NAME is runing..."
+        else
+            echo "$APP_NAME is not runing..."
+        fi
     else
         echo "$APP_NAME is not runing..."
     fi
